Mise à jour immédiate de l'état local lors de l'envoi d'une commande. Ajout d'un rechargement de sécurité si l'API renvoie une erreur. Ajustement du Spinner

+++ app/src/main/java/com/robin/polyhome/DeviceData.kt
@@ -4,8 +4,8 @@ data class DeviceData(
     val id: String,
     val type: String,
     val availableCommands: ArrayList<String>,
-    var power: Int?,
-    var opening: Int?
+    var power: Double?,
+    var opening: Double?
 )
 
 data class DevicesResponse(
